fix: refactor dnd comp
