+++ src/lib/layout/network-menu/NetworkAccordionPinned.tsx
@@ -1,0 +1,138 @@
+import {
+  AccordionButton,
+  AccordionIcon,
+  AccordionItem,
+  AccordionPanel,
+  Flex,
+  Heading,
+} from "@chakra-ui/react";
+import type { Active, DragEndEvent } from "@dnd-kit/core";
+import {
+  closestCenter,
+  defaultDropAnimationSideEffects,
+  DndContext,
+  DragOverlay,
+  MouseSensor,
+  TouchSensor,
+  useSensor,
+  useSensors,
+} from "@dnd-kit/core";
+import {
+  SortableContext,
+  verticalListSortingStrategy,
+} from "@dnd-kit/sortable";
+import { useMemo, useState } from "react";
+
+import { useCelatoneApp } from "lib/app-provider";
+import { useNetworkStore } from "lib/providers/store";
+import type { Network } from "lib/stores/networks";
+import type { Nullable, Option } from "lib/types";
+
+import { NetworkCardDraggable } from "./NetworkCardDraggable";
+
+interface NetworkAccodionPinnedProps {
+  pinnedNetworks: Network[];
+  cursor: Option<number>;
+  setCursor: (index: Option<number>) => void;
+  onClose: () => void;
+}
+
+export const NetworkAccodionPinned = ({
+  pinnedNetworks,
+  cursor,
+  setCursor,
+  onClose,
+}: NetworkAccodionPinnedProps) => {
+  const { currentChainId } = useCelatoneApp();
+  const { setPinnedNetworks } = useNetworkStore();
+  const [dndActive, setDndActive] = useState<Nullable<Active>>(null);
+
+  // Drag and drop feature
+  const activeItem = useMemo(
+    () => pinnedNetworks.find((item) => item.id === dndActive?.id),
+    [dndActive, pinnedNetworks]
+  );
+
+  const sensors = useSensors(
+    useSensor(MouseSensor, {
+      activationConstraint: {
+        delay: 100,
+        tolerance: 5,
+      },
+    }),
+    useSensor(TouchSensor)
+  );
+
+  const handleDragEnd = (event: DragEndEvent) => {
+    const { active, over } = event;
+
+    if (active.id !== over?.id) {
+      setPinnedNetworks(active.id.toString(), over?.id.toString());
+    }
+  };
+
+  return (
+    <AccordionItem hidden={pinnedNetworks.length === 0}>
+      <AccordionButton p={0}>
+        <Flex mb={4} justifyContent="space-between" w="full">
+          <Heading as="h6" variant="h6">
+            Pinned Network
+          </Heading>
+          <AccordionIcon color="gray.600" />
+        </Flex>
+      </AccordionButton>
+      <AccordionPanel p={0} mb={2}>
+        <DndContext
+          sensors={sensors}
+          collisionDetection={closestCenter}
+          onDragStart={({ active }) => {
+            setDndActive(active);
+          }}
+          onDragEnd={handleDragEnd}
+          onDragCancel={() => {
+            setDndActive(null);
+          }}
+        >
+          <SortableContext
+            items={pinnedNetworks.map((item) => item.chainId)}
+            strategy={verticalListSortingStrategy}
+          >
+            {pinnedNetworks.map((item, index) => (
+              <NetworkCardDraggable
+                key={item.chainId}
+                chainId={item.chainId}
+                isSelected={item.chainId === currentChainId}
+                index={index}
+                cursor={cursor}
+                setCursor={setCursor}
+                onClose={onClose}
+              />
+            ))}
+          </SortableContext>
+          <DragOverlay
+            dropAnimation={{
+              sideEffects: defaultDropAnimationSideEffects({
+                styles: {
+                  active: {
+                    opacity: "0.4",
+                  },
+                },
+              }),
+            }}
+          >
+            {activeItem ? (
+              <NetworkCardDraggable
+                key={activeItem.chainId}
+                chainId={activeItem.chainId}
+                isSelected={activeItem.chainId === currentChainId}
+                cursor={cursor}
+                setCursor={setCursor}
+                onClose={onClose}
+              />
+            ) : null}
+          </DragOverlay>
+        </DndContext>
+      </AccordionPanel>
+    </AccordionItem>
+  );
+};
